Добавлены методы ответа клиенту
