+++ src/main/java/Main.java
@@ -1,37 +1,35 @@
+import org.apache.http.client.utils.URLEncodedUtils;
+
+import java.io.BufferedOutputStream;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
 public class Main {
+
+
     public static void main(String[] args) {
         Server server = new Server();
+        server.addHandler("GET", "/index.html", (request, responseStream) -> {
+            Request.successfulRequest(request, responseStream);
+            Files.copy(Path.of(".", "public", request.getPath()), responseStream);
+            responseStream.flush();
+
+        });
+
         server.addHandler("GET", "/classic.html", (request, responseStream) -> {
-            responseStream.write((
-                    "HTTP/1.1 200 OK\r\n" +
-                            "Content-Type: " + request.getType() + "\r\n" +
-                            "Content-Length: " + request.getLength() + "\r\n" +
-                            "Connection: close\r\n" +
-                            "\r\n"
-            ).getBytes());
+            Request.timeRequest(request, responseStream);
             Files.copy(Path.of(".", "public", request.getPath()), responseStream);
             responseStream.flush();
 
         });
 
         server.addHandler("POST", "message", (request, responseStream) -> {
-            responseStream.write((
-                    "HTTP/1.1 200 OK\r\n" +
-                            "Content-Type: " + request.getType() + "\r\n" +
-                            "Content-Length: " + request.getLength() + "\r\n" +
-                            "Connection: close\r\n" +
-                            "\r\n"
-            ).getBytes());
+            Request.successfulRequest(request, responseStream);
             Files.copy(Path.of(".", "public", request.getPath()), responseStream);
             responseStream.flush();
 
         });
-
-
-
 
         server.start(9999);
