+++ src/main/java/Request.java
@@ -1,6 +1,8 @@
+import java.io.BufferedOutputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.time.LocalDateTime;
 import java.util.List;
 
 public class Request {
@@ -17,6 +19,44 @@ public class Request {
         this.headers = headers;
 
         this.path = path;
+    }
+
+    public static void successfulRequest(Request request, BufferedOutputStream out) throws IOException {
+        out.write((
+                "HTTP/1.1 200 OK\r\n" +
+                        "Content-Type: " + request.getType() + "\r\n" +
+                        "Content-Length: " + request.getLength() + "\r\n" +
+                        "Connection: close\r\n" +
+                        "\r\n"
+        ).getBytes());
+
+    }
+    public static void badRequest(BufferedOutputStream out) throws IOException {
+        out.write((
+                "HTTP/1.1 404 Not Found\r\n" +
+                        "Content-Length: 0\r\n" +
+                        "Connection: close\r\n" +
+                        "\r\n"
+        ).getBytes());
+        out.flush();
+    }
+
+    public static void timeRequest(Request request, BufferedOutputStream out) throws IOException {
+        var filePath = Path.of(".", "public", request.getPath());
+        final var template = Files.readString(filePath);
+        final var content = template.replace(
+                "{time}",
+                LocalDateTime.now().toString()
+        ).getBytes();
+        out.write((
+                "HTTP/1.1 200 OK\r\n" +
+                        "Content-Type: " + request.getType() + "\r\n" +
+                        "Content-Length: " + content.length + "\r\n" +
+                        "Connection: close\r\n" +
+                        "\r\n"
+        ).getBytes());
+        out.write(content);
+        out.flush();
     }
